+++ aurmr_tasks/src/aurmr_tasks/util.py
@@ -8,6 +8,8 @@ from geometry_msgs.msg import PoseStamped
 from moveit_commander.conversions import pose_to_list
 from smach import State, StateMachine
 
+from aurmr_perception.util import qv_mult, quat_msg_to_vec
+
 
 def apply_offset_to_pose(pose, offset, offset_frame=None, tf_buffer=None):
     if not isinstance(pose, PoseStamped):
@@ -23,6 +25,16 @@ def apply_offset_to_pose(pose, offset, offset_frame=None, tf_buffer=None):
     # It would be better to grab the transform once and apply it to the offset vector
     offset_pose = tf_buffer.transform(offset_pose, pose.header.frame_id, rospy.Duration(1))
     return offset_pose
+
+def add_offset(offset, grasp_pose):
+    v = qv_mult(
+        quat_msg_to_vec(grasp_pose.pose.orientation), (0, 0, offset))
+    offset_pose = deepcopy(grasp_pose)
+    offset_pose.pose.position.x += v[0]
+    offset_pose.pose.position.y += v[1]
+    offset_pose.pose.position.z += v[2]
+    return offset_pose
+
 
 
 class Formulator(State):
